feat: Refactor ConversationsTable directory
